rest-server/requirements.txt
Flask==2.1.0
gunicorn==20.1.0
cloud-sql-python-connector==1.2.2
SQLAlchemy==2.0.9
pymysql==1.0.3
GitPython==3.1.31
protobuf==4.22.3
Requests==2.28.2